+++ packages/lib/src/types/objectBased/typesModel.ts
@@ -153,6 +153,8 @@ export interface ModelTypeInfoProps {
  * `types.model` type info.
  */
 export class ModelTypeInfo extends TypeInfo {
+  readonly kind = "model"
+
   private _props = lazy(() => {
     const objSchema = getInternalModelClassPropsInfo(this.modelClass)
 
@@ -189,6 +191,32 @@ export class ModelTypeInfo extends TypeInfo {
 
   get props(): ModelTypeInfoProps {
     return this._props()
+  }
+
+  override isTopLevelPropertyContainer(): boolean {
+    return true
+  }
+
+  override getTopLevelPropertyTypeInfo(propertyName: string): TypeInfo | undefined {
+    return this.props[propertyName]?.typeInfo
+  }
+
+  override shouldTraverseChildrenAfterTopLevelPropertySelection(): boolean {
+    return false
+  }
+
+  override findChildTypeInfo(
+    predicate: (childTypeInfo: TypeInfo) => boolean
+  ): TypeInfo | undefined {
+    const props = this.props
+    const propNames = Object.keys(props)
+    for (let i = 0; i < propNames.length; i++) {
+      const childTypeInfo = props[propNames[i]].typeInfo
+      if (childTypeInfo && predicate(childTypeInfo)) {
+        return childTypeInfo
+      }
+    }
+    return undefined
   }
 
   get modelType(): string {